refactor: modularize tool definitions by model family
